refactor(cli): cleanups to new module graph (#7846)
