+++ cli/main.rs
@@ -25,7 +25,6 @@ pub mod fmt_errors;
 mod fs;
 pub mod global_state;
 mod global_timer;
-mod graph;
 pub mod http_cache;
 mod http_util;
 mod import_map;
@@ -38,6 +37,7 @@ mod lockfile;
 mod media_type;
 mod metrics;
 mod module_graph;
+mod module_graph2;
 mod op_fetch_asset;
 pub mod ops;
 pub mod permissions;
